Fix uprefix issue in solr search
setting uprefix was also setting the name of the text field to have the
prefix as well.  Removed the setuprefix call and am now setting the
prefix values for the fields by hand using the
Utils:createDocumentFieldName which is the same logic that is being
used by the search service.